feat(GameManager): add GameModes and a restart method
Now the game has two modes, the Confined one keep the ships confined in the camera area. The Free mode lets the ships pass the camera bounds but theyre warped to the opposite edge.

Also add a restart method and a method to reset all variables on Awake, so the restart truly returns the game to its initial state and doesnt generate any bugs.